Ref: В store вместо отфильтрованного списка фильмов хранится исходный (неизменяемый) список, соотв. удален лишний код
