+++ src/components/blocks/movies-list/movies-list.jsx
@@ -1,5 +1,4 @@
 import React, {useState} from "react";
-import {connect} from 'react-redux';
 import PropTypes from "prop-types";
 import MovieProp from '../../props/movie.prop';
 import MovieCard from "../../blocks/movie-card/movie-card";
@@ -41,10 +40,4 @@ MoviesList.propTypes = {
   listSize: PropTypes.number.isRequired,
 };
 
-
-const mapStateToProps = (state) => ({
-  films: state.films,
-});
-
-export {MoviesList};
-export default connect(mapStateToProps, null)(MoviesList);
+export default MoviesList;
