+++ src/store/selectors.js
@@ -1,0 +1,6 @@
+export const getFilmsByGenre = (genre, films) => {
+  if (genre) {
+    return films.filter((film) => film.genre === genre);
+  }
+  return films;
+};
